requirements.txt
# --- Core FastAPI Web App ---
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
requests

# --- Authentication ---
PyJWT
bcrypt

# --- MongoDB ---
pymongo
motor

# --- Web Scraping & Parsing ---
trafilatura
playwright
beautifulsoup4

# --- PDF Document Processing ---
langchain-core==0.3.72
langchain_community==0.3.27
pypdf==5.9.0

# --- LangChain RAG Components ---
langchain-groq==0.3.6
langchain-pinecone==0.2.11

# --- Pinecone Vector DB ---
pinecone==7.3.0

# --- Embeddings & Tokenization ---
tiktoken==0.9.0
sentence-transformers==5.0.0

# --- Validation & Settings ---
pydantic==2.9.2
pydantic-settings==2.8.1

# --- Visualization ---
matplotlib

# --- Excel Export ---
pandas
openpyxl